+++ examples/data-sources/kilhog_network/data-source.tf
@@ -1,0 +1,3 @@
+data "kilhog_network" "production" {
+  name = "production"
+}