+++ examples/data-sources/kilhog_subnet/data-source.tf
@@ -1,0 +1,8 @@
+data "kilhog_network" "production" {
+  name = "production"
+}
+
+data "kilhog_subnet" "dmz" {
+  network_id = data.kilhog_network.production.id
+  name       = "dmz"
+}